+++ packages/jimp/package.json
@@ -63,7 +63,8 @@
     "@jimp/png": "0.3.5",
     "@jimp/bmp": "0.3.5",
     "@jimp/tiff": "0.3.5",
-    "@jimp/plugin-resize": "0.3.5"
+    "@jimp/plugin-resize": "0.3.5",
+    "@jimp/plugin-blit": "0.3.5"
   },
   "devDependencies": {
     "babelify": "^9.0.0",
