+++ packages/plugin-blit/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-blit
+
+Blit an image.
